+++ packages/iov-ledger-bns/src/statetracker.spec.ts
@@ -12,6 +12,20 @@ describe("StateTracker", () => {
     expect(tracker.state.value).toEqual(LedgerState.Disconnected);
   });
 
+  it("has running state false when created", () => {
+    const tracker = new StateTracker();
+    expect(tracker.running).toEqual(false);
+  });
+
+  it("has running state true when started and false when stopped", () => {
+    const tracker = new StateTracker();
+    expect(tracker.running).toEqual(false);
+    tracker.start();
+    expect(tracker.running).toEqual(true);
+    tracker.stop();
+    expect(tracker.running).toEqual(false);
+  });
+
   it("changes state to app open", done => {
     pendingWithoutLedger();
 
@@ -20,6 +34,7 @@ describe("StateTracker", () => {
     tracker.state.updates.subscribe({
       next: value => {
         if (value === LedgerState.IovAppOpen) {
+          tracker.stop();
           done();
         }
       },
@@ -27,5 +42,19 @@ describe("StateTracker", () => {
       complete: fail,
     });
     tracker.start();
+  });
+
+  it("changes to disconnected when tracking is stopped", async () => {
+    pendingWithoutLedger();
+
+    const tracker = new StateTracker();
+    expect(tracker.state.value).toEqual(LedgerState.Disconnected);
+
+    tracker.start();
+    await tracker.state.waitFor(LedgerState.IovAppOpen);
+    expect(tracker.state.value).toEqual(LedgerState.IovAppOpen);
+
+    tracker.stop();
+    expect(tracker.state.value).toEqual(LedgerState.Disconnected);
   });
 });