+++ packages/iov-ledger-bns/src/statetracker.ts
@@ -1,8 +1,8 @@
 import { DefaultValueProducer, ValueAndUpdates } from "@iov/keycontrol";
-import TransportNodeHid, { DescriptorEvent } from "@ledgerhq/hw-transport-node-hid";
+import TransportNodeHid, { DescriptorEvent, Subscription } from "@ledgerhq/hw-transport-node-hid";
 
 import { appVersion } from "./app";
-import { connectToFirstLedger } from "./exchange";
+import { connectToFirstLedger, LedgerErrorResponse } from "./exchange";
 
 export enum LedgerState {
   Disconnected,
@@ -12,27 +12,39 @@ export enum LedgerState {
 
 export class StateTracker {
   private static async checkConectedAndAppOpen(): Promise<LedgerState> {
+    // tslint:disable-next-line:no-let
+    let transport: TransportNodeHid | undefined;
     try {
-      const transport = await connectToFirstLedger();
-      try {
-        // use appVersion() as a status check: if it works, we are in the app
-        // otherwise no
-        await appVersion(transport);
-        // console.log(`>>> Entered app (version ${version})`);
-        return LedgerState.IovAppOpen;
-      } catch (_) {
-        // not in app
-        return LedgerState.Connected;
-      }
+      transport = await connectToFirstLedger();
     } catch (_) {
       // console.log("Error connecting to ledger: " + err);
       return LedgerState.Disconnected;
+    }
+
+    try {
+      // use appVersion() as a status check: if it works, we are in the app
+      // otherwise no
+      await appVersion(transport);
+      return LedgerState.IovAppOpen;
+    } catch (error) {
+      if (error instanceof LedgerErrorResponse && error.code === 0x6e00) {
+        // not in app
+        return LedgerState.Connected;
+      } else {
+        throw error;
+      }
     }
   }
 
   public readonly state: ValueAndUpdates<LedgerState>;
 
+  public get running(): boolean {
+    return this.listeningSubscription !== undefined;
+  }
+
   private readonly stateProducer: DefaultValueProducer<LedgerState>;
+  // tslint:disable-next-line:readonly-keyword
+  private listeningSubscription: Subscription | undefined;
 
   constructor() {
     this.stateProducer = new DefaultValueProducer(LedgerState.Disconnected);
@@ -40,7 +52,8 @@ export class StateTracker {
   }
 
   public start(): void {
-    TransportNodeHid.listen({
+    // tslint:disable-next-line:no-object-mutation
+    this.listeningSubscription = TransportNodeHid.listen({
       next: e => this.handleEvent(e),
       error: e => {
         throw e;
@@ -49,6 +62,15 @@ export class StateTracker {
         throw new Error("TransportNodeHid.listen completed. What does that mean?");
       },
     });
+  }
+
+  public stop(): void {
+    if (this.listeningSubscription) {
+      this.listeningSubscription.unsubscribe();
+      // tslint:disable-next-line:no-object-mutation
+      this.listeningSubscription = undefined;
+    }
+    this.stateProducer.update(LedgerState.Disconnected);
   }
 
   /**