+++ packages/iov-ledger-bns/types/statetracker.d.ts
@@ -7,8 +7,11 @@ export declare enum LedgerState {
 export declare class StateTracker {
     private static checkConectedAndAppOpen;
     readonly state: ValueAndUpdates<LedgerState>;
+    readonly running: boolean;
     private readonly stateProducer;
+    private listeningSubscription;
     constructor();
     start(): void;
+    stop(): void;
     private handleEvent;
 }